+++ arc/context/autocomplete.py
@@ -51,7 +51,7 @@ class AutocompleteData(t.Generic[ClientT, ChoiceT]):
         """
         if self.focused_option is None:
             return None
-        return t.cast(ChoiceT | str, self.focused_option.value)
+        return t.cast("ChoiceT | str", self.focused_option.value)
 
     @property
     def guild_id(self) -> hikari.Snowflake | None:
@@ -69,17 +69,31 @@ class AutocompleteData(t.Generic[ClientT, ChoiceT]):
         return self.interaction.member
 
     @property
+    def invocation_context(self) -> hikari.ApplicationContextType:
+        """The context in which the interaction was invoked."""
+        return self.interaction.context
+
+    @property
+    def authorizing_integration_owners(self) -> t.Mapping[hikari.ApplicationIntegrationType, hikari.Snowflake]:
+        """Includes details about the authorizing user or server for the installation(s) relevant to the interaction.
+
+        For apps installed to a user, it can be used to tell the difference between the authorizing user
+        and the user that triggered an interaction (like a message component).
+        """
+        return self.interaction.authorizing_integration_owners
+
+    @property
     def user(self) -> hikari.User:
         """The user that triggered the interaction."""
         return self.interaction.user
 
+    @property
+    def channel(self) -> hikari.InteractionChannel:
+        return self.interaction.channel
+
     def get_guild(self) -> hikari.Guild | None:
         """Get the guild that triggered the interaction."""
         return self.interaction.get_guild()
-
-    def get_channel(self) -> hikari.TextableGuildChannel | None:
-        """Get the channel that triggered the interaction."""
-        return self.interaction.get_channel()
 
 
 # MIT License